% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{new_process}
\alias{new_process}
\title{Create an object representing a tumorigenic process}
\usage{
new_process(
  name,
  description,
  class,
  icon,
  flip_vertical,
  flip_horizontal,
  scars
)
}
\arguments{
\item{name}{The name of the tumorigenic process. Must be one of the values
returned by \code{\link[=valid_processes]{valid_processes()}} (string).}

\item{description}{A detailed description of the tumorigenic process (string).}

\item{class}{The classification of the process, either \strong{"endogenous"}
(originating from internal factors) or \strong{"exogenous"} (originating from external factors) (string).}

\item{icon}{A string identifier for the favicon icon representing the process
(e.g., \code{faRadiation} for radiation processes) (string).}

\item{flip_vertical}{Logical flag indicating if the icon should be flipped vertically (boolean).}

\item{flip_horizontal}{Logical flag indicating if the icon should be flipped horizontally (boolean).}

\item{scars}{A list of \code{scar} objects associated with this tumorigenic process.
Each scar must have the same \code{process} name as the one provided here (list).}
}
\value{
An object of class \strong{process} with the following properties:
\describe{
\item{name}{The name of the tumorigenic process (string).}
\item{description}{A detailed description of the tumorigenic process (string).}
\item{icon}{The identifier for the icon used to visually represent the process (string).}
\item{flip_vertical}{A logical flag indicating whether the icon should be flipped vertically (boolean).}
\item{flip_horizontal}{A logical flag indicating whether the icon should be flipped horizontally (boolean).}
\item{dna}{Logical value indicating if this process leaves scars on DNA (boolean).}
\item{rna}{Logical value indicating if this process leaves scars on RNA (boolean).}
\item{meth}{Logical value indicating if this process leaves scars on methylation markers (boolean).}
\item{scars}{A list of \code{scar} objects associated with the process. Each \code{scar} must have the same \code{process} name (list).}
}
}
\description{
This function constructs a \code{process} object representing a tumorigenic process
(e.g., endogenous or exogenous processes like radiation or genetic mutations).
The object includes details about the process, such as its name, description,
icon, scars associated with the process, and whether the process affects DNA, RNA,
or methylation markers.
}
\examples{
# Creating two scars for Radiotherapy
high_indel_burden <- new_scar(
  process = "Radiotherapy",
  scar_name = "High Indel Burden",
  marker_of = "Radiation Induced Damage",
  disease_specificity = "Pan-Cancer",
  diseases = "cancer",
  modality = "DNA",
  measurement = "INDEL / SBS ratio",
  description = "High INDEL counts (mostly deletions)",
  experiment = "Analysis of 12 radiation-associated tumors compared to radiation-naive tumors",
  paper_url = "https://www.nature.com/articles/ncomms12605",
  specificity = "Low",
  specificity_tooltip = "Also observed in BRCA1/BRCA2 breast cancer.",
  sensitivity = "Moderate",
  sensitivity_tooltip = "Above median in all radiation-induced tumors",
  tool_name = "raDNA",
  tool_url = "https://github.com/selkamand/radna"
)

high_deletion_burden <- new_scar(
  process = "Radiotherapy",
  scar_name = "High Deletion Burden",
  marker_of = "Radiation Induced Damage",
  disease_specificity = "Pan-Cancer",
  diseases = "cancer",
  modality = "DNA",
  measurement = "Deletion / Insertion ratio",
  description = "Small deletions in regions of microhomology.",
  experiment = "Analysis of 12 radiation-associated tumors",
  paper_url = "https://www.nature.com/articles/ncomms12605",
  specificity = "Low",
  specificity_tooltip = "Also observed in BRCA1/BRCA2 breast cancer.",
  sensitivity = "Moderate",
  sensitivity_tooltip = "Above median in 11/12 radiation-induced tumors.",
  tool_name = "raDNA",
  tool_url = "https://github.com/selkamand/radna"
)

# Create the Radiotherapy process object with two scars
radiotherapy_process <- new_process(
  name = "Radiotherapy",
  description = "Damage caused by ionizing radiation therapy",
  class = "exogenous",
  icon = "faRadiation",
  flip_vertical = FALSE,
  flip_horizontal = TRUE,
  scars = list(high_indel_burden, high_deletion_burden)
)
}
