% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{new_scar}
\alias{new_scar}
\title{Create a new scar created by tumorigenic processes}
\usage{
new_scar(
  process,
  scar_name,
  marker_of,
  disease_specificity,
  diseases,
  modality,
  measurement,
  description,
  experiment,
  paper_url,
  specificity,
  specificity_tooltip,
  sensitivity,
  sensitivity_tooltip,
  tool_name,
  tool_url
)
}
\arguments{
\item{process}{The tumorigenic process responsible for this scar (e.g., radiation or molecular dysfunction).
Must be one of the values returned by \code{\link[=valid_processes]{valid_processes()}} (string).}

\item{scar_name}{The name of the scar (string).}

\item{marker_of}{A detailed description of what the scar is a biomarker for,
such as a specific biological process or mutation signature (string).}

\item{disease_specificity}{Specifies if the scar is associated with \strong{"Pan-Cancer"}
or \strong{"Disease Specific"} conditions (string).}

\item{diseases}{A character vector of diseases linked to the scar. If \code{disease_specificity = "Pan-Cancer"},
this must be \strong{"cancer"}.}

\item{modality}{The biological modality in which the scar is detected. Must be one of \strong{"DNA"},
\strong{"RNA"}, or \strong{"Methylation"} (string).}

\item{measurement}{The specific metric used to detect or quantify the scar (string).}

\item{description}{A plain-language description that explains how this scar is related to the
tumorigenic process or biological phenomenon it marks (string).}

\item{experiment}{A plain-language description of the experiment or analysis that discovered this scar
and established its association with the tumorigenic process (string).}

\item{paper_url}{The URL of the scientific publication that provides evidence for this scar's existence and relevance (string).}

\item{specificity}{Describes how specifically the scar is associated with the process it marks.
One of \strong{"High"}, \strong{"Moderate"}, \strong{"Low"}, or \strong{"Unknown"} (string).}

\item{specificity_tooltip}{A description explaining how the specificity was inferred,
including whether other processes might also contribute to the appearance of this scar (string).}

\item{sensitivity}{Indicates how sensitive the scar is to the process it marks.
One of \strong{"High"}, \strong{"Moderate"}, \strong{"Low"}, or \strong{"Unknown"} (string).}

\item{sensitivity_tooltip}{A description explaining how the sensitivity was inferred (string).}

\item{tool_name}{The name of the computational or laboratory tool used for scar detection, or \code{NULL}
if no tool is available (string or \code{NULL}).}

\item{tool_url}{The URL for the tool used for scar detection, or \code{NULL} if no tool is available (string or \code{NULL}).}
}
\value{
An object of class \strong{scar}, representing the tumorigenic process and its associated biomarker. Each \code{scar} object contains the following properties:
\describe{
\item{scar_name}{The name of the scar (string).}
\item{process}{The tumorigenic process that generates the scar (string).}
\item{marker_of}{A description of the biological phenomenon the scar serves as a biomarker of (string).}
\item{disease_specificity}{Whether the scar is \strong{"Pan-Cancer"} or \strong{"Disease Specific"} (string).}
\item{diseases}{A character vector of diseases associated with the scar (character).}
\item{modality}{The biological modality in which the scar is detected (e.g., DNA, RNA, Methylation) (string).}
\item{measurement}{The metric used to detect or quantify the scar (string).}
\item{description}{A description explaining how the scar relates to the process (string).}
\item{experiment}{A description of the experiment that identified the scar and its association to the process (string).}
\item{paper_url}{The URL of the publication detailing the scar (string).}
\item{specificity}{How specific the scar is to the process, with possible values \strong{"High"}, \strong{"Moderate"}, \strong{"Low"}, or \strong{"Unknown"} (string).}
\item{specificity_tooltip}{An explanation of how the specificity of the scar was determined (string).}
\item{sensitivity}{How sensitive the scar is as a biomarker of the process, with possible values \strong{"High"}, \strong{"Moderate"}, \strong{"Low"}, or \strong{"Unknown"} (string).}
\item{sensitivity_tooltip}{An explanation of how the sensitivity of the scar was determined (string).}
\item{tool_name}{The name of the tool used to detect the scar, if available (string or \code{NULL}).}
\item{tool_url}{The URL of the tool used for scar detection, if available (string or \code{NULL}).}
}
}
\description{
This function constructs a new \code{scar} object that represents a characteristic pattern observed in DNA,RNA or Methylation
that is produced as a by-product of tumorigenic processes (e.g. radiation exposure or molecular dysfunctions).
}
\examples{
new_scar(
  process = "Radiotherapy",
  scar_name = "High Indel Burden",
  marker_of = "Radiation Induced Damage",
  disease_specificity = "Pan-Cancer",
  diseases = "cancer",
  modality = "DNA",
  measurement = "INDEL / SBS ratio",
  description = "Higher number of INDELS (mainly deletions)",
  experiment = "Analysis of 12 radiation-associated tumors compared to radiation-naive tumors",
  paper_url = "https://www.nature.com/articles/ncomms12605",
  specificity = "Low",
  specificity_tooltip = "Also observed in BRCA1/BRCA2 breast cancer.",
  sensitivity = "Moderate",
  sensitivity_tooltip = "Above median in all radiation-induced tumours.",
  tool_name = "raDNA",
  tool_url = "https://github.com/selkamand/radna"
)

}
