% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{print.process}
\alias{print.process}
\title{Print process}
\usage{
\method{print}{process}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{added for uniformity with generic. Does not do anything when printing process objects}
}
\description{
Print process
}
