% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{valid_processes}
\alias{valid_processes}
\title{Retrieve a list of valid tumorigenic processes}
\usage{
valid_processes()
}
\value{
A character vector of valid tumorigenic processes.
}
\description{
This function returns a character vector of valid tumorigenic processes
that can be associated with a scar. These processes are high-level descriptions
of biological or treatment-related phenomena that leave characteristic
mutation signatures or multiomic scars in cancer tissues.
}
\examples{
valid_processes()
}
